Split of the JitCompareScreen into its own files
