+++ UI/UI.vcxproj.filters
@@ -95,6 +95,9 @@
     <ClCompile Include="DriverManagerScreen.cpp">
       <Filter>Screens</Filter>
     </ClCompile>
+    <ClCompile Include="JitCompareScreen.cpp">
+      <Filter>Screens</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GameInfoCache.h" />
@@ -188,6 +191,9 @@
       <Filter>Screens</Filter>
     </ClInclude>
     <ClInclude Include="DriverManagerScreen.h">
+      <Filter>Screens</Filter>
+    </ClInclude>
+    <ClInclude Include="JitCompareScreen.h">
       <Filter>Screens</Filter>
     </ClInclude>
   </ItemGroup>
