+++ UWP/UI_UWP/UI_UWP.vcxproj.filters
@@ -38,6 +38,7 @@
     <ClCompile Include="..\..\UI\RetroAchievementScreens.cpp" />
     <ClCompile Include="..\..\UI\DebugOverlay.cpp" />
     <ClCompile Include="..\..\UI\DriverManagerScreen.cpp" />
+    <ClCompile Include="..\..\UI\JitCompareScreen.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
@@ -77,5 +78,6 @@
     <ClInclude Include="..\..\UI\RetroAchievementScreens.h" />
     <ClInclude Include="..\..\UI\DebugOverlay.h" />
     <ClInclude Include="..\..\UI\DriverManagerScreen.h" />
+    <ClInclude Include="..\..\UI\JitCompareScreen.h" />
   </ItemGroup>
 </Project>
